Add capability checks to non-REST write operations, harden input
Security audit findings:

- PatchBlockAbility: add edit_post check before wp_update_post
- DuplicatePostAbility: add edit_post + create_posts checks
- ManageRevisionsAbility: add edit_post check before restore
- CreateTranslationAbility: add edit_post + create_posts checks
- CreateTermTranslationAbility: add edit_terms check
- UpdateStringTranslationAbility: require manage_options
- ManageRedirectsAbility: require publish_pages for create action
- BulkUpdatePostsAbility: block protected meta keys via is_protected_meta()
- Cast all user-provided IDs to (int) across all abilities

REST-delegating abilities (content CRUD, terms) already enforce
capabilities via rest_do_request() — the REST controller's own
permission_callback runs. These are not affected.

12 security tests verify that subscribers cannot:
- patch blocks, duplicate posts, restore revisions
- create translations, create redirects
- write protected meta keys

+++ src/Abilities/GenericPostTypeAbility.php
@@ -207,7 +207,7 @@ final class GenericPostTypeAbility
         if (! $route) {
             return new WP_Error('invalid_type', 'Unknown content type: '.($input['type'] ?? ''));
         }
-        $id = $input['id'] ?? 0;
+        $id = (int) ($input['id'] ?? 0);
         unset($input['type'], $input['id']);
 
         $response = self::restGet("{$route}/{$id}", $input);
@@ -252,7 +252,7 @@ final class GenericPostTypeAbility
         if (! $route) {
             return new WP_Error('invalid_type', 'Unknown content type: '.($input['type'] ?? ''));
         }
-        $id = $input['id'] ?? 0;
+        $id = (int) ($input['id'] ?? 0);
 
         $acfFields = $input['fields'] ?? null;
         unset($input['type'], $input['id'], $input['fields']);
@@ -279,7 +279,7 @@ final class GenericPostTypeAbility
         if (! $route) {
             return new WP_Error('invalid_type', 'Unknown content type: '.($input['type'] ?? ''));
         }
-        $id = $input['id'] ?? 0;
+        $id = (int) ($input['id'] ?? 0);
         $force = $input['force'] ?? false;
 
         $request = new \WP_REST_Request('DELETE', "{$route}/{$id}");
